change deleteProductItemById

+++ src/products/products.service.js
@@ -1,4 +1,6 @@
 import { ProductItem } from './products.model.js'
+import path from 'path'
+import fs from 'fs'
 
 export async function fetchProductItems() {
   return await ProductItem.find()
@@ -10,5 +12,20 @@ export async function createProductItem(data) {
 }
 
 export async function deleteProductItemById(customId) {
+  const product = await ProductItem.findOne({ id: customId })
+  if (!product) return null
+
+  if (product.image) {
+    const imagePath = path.join(process.cwd(), product.image)
+
+    try {
+      if (fs.existsSync(imagePath)) {
+        fs.unlinkSync(imagePath)
+      }
+    } catch (err) {
+      console.error('Failed to delete image file:', err)
+    }
+  }
+
   return await ProductItem.findOneAndDelete({ id: customId })
 }
